Move visualization settings from viz footer to editor and restyle to match space efficient qb
