+++ resources/frontend_client/app/query_builder/run_button.react.js
@@ -7,16 +7,13 @@ export default React.createClass({
         isRunning: React.PropTypes.bool.isRequired,
         runFn: React.PropTypes.func.isRequired
     },
-    run: function() {
-
-    },
     render: function () {
         // default state is to not render anything if we can't actually run
         var runButton = false;
         if (this.props.canRun) {
-            var runButtonText = (this.props.isRunning) ? "Loading..." : "Find out!";
+            var runButtonText = (this.props.isRunning) ? "Loading..." : "Run updated query";
             runButton = (
-                <button className="Button Button--primary" onClick={this.props.runFn}>{runButtonText}</button>
+                <button className="Button Button--primary circular" onClick={this.props.runFn}>{runButtonText}</button>
             );
         }
