+++ resources/frontend_client/app/query_builder/visualization.react.js
@@ -1,8 +1,6 @@
 'use strict';
 
 import { CardRenderer } from '../card/card.charting';
-import Icon from './icon.react';
-import PopoverWithTrigger from './popover_with_trigger.react';
 import QueryVisualizationTable from './visualization_table.react';
 import QueryVisualizationChart from './visualization_chart.react';
 import QueryVisualizationObjectDetailTable from './visualization_object_detail_table.react';
@@ -16,40 +14,15 @@ export default React.createClass({
         visualizationSettingsApi: React.PropTypes.object.isRequired,
         card: React.PropTypes.object.isRequired,
         result: React.PropTypes.object,
-        setDisplayFn: React.PropTypes.func.isRequired,
-        setChartColorFn: React.PropTypes.func.isRequired,
         setSortFn: React.PropTypes.func.isRequired,
         cellIsClickableFn: React.PropTypes.func,
         cellClickedFn: React.PropTypes.func
     },
 
-    visualizationTypeNames: {
-        'scalar':  { displayName: 'Number',      iconName: 'number' },
-        'table':   { displayName: 'Table',       iconName: 'table' },
-        'line':    { displayName: 'Line',        iconName: 'line' },
-        'bar':     { displayName: 'Bar',         iconName: 'bar' },
-        'pie':     { displayName: 'Pie',         iconName: 'pie' },
-        'area':    { displayName: 'Area',        iconName: 'area' },
-        'state':   { displayName: 'State map',   iconName: 'statemap' },
-        'country': { displayName: 'Country map', iconName: 'countrymap' },
-        'pin_map': { displayName: 'Pin map',     iconName: 'pinmap' }
-    },
-
     getDefaultProps: function() {
         return {
             // NOTE: this should be more dynamic from the backend, it's set based on the query lang
-            maxTableRows: 2000,
-            visualizationTypes: [
-                'scalar',
-                'table',
-                'line',
-                'bar',
-                'pie',
-                'area',
-                'state',
-                'country',
-                'pin_map'
-            ]
+            maxTableRows: 2000
         };
     },
 
@@ -73,173 +46,18 @@ export default React.createClass({
         return (JSON.stringify(this.props.card.dataset_query) !== this.state.origQuery);
     },
 
-    hasLatitudeAndLongitudeColumns: function(columnDefs) {
-        var hasLatitude = false,
-            hasLongitude = false;
-        columnDefs.forEach(function(col, index) {
-            if (col.special_type &&
-                    col.special_type === "latitude") {
-                hasLatitude = true;
-
-            } else if (col.special_type &&
-                    col.special_type === "longitude") {
-                hasLongitude = true;
-            }
-        });
-
-        return (hasLatitude && hasLongitude);
-    },
-
-    isSensibleChartDisplay: function(display) {
-        var data = (this.props.result) ? this.props.result.data : null;
-        switch (display) {
-            case "table":
-                // table is always appropriate
-                return true;
-            case "scalar":
-                // a 1x1 data set is appropriate for a scalar
-                return (data && data.rows && data.rows.length === 1 && data.cols && data.cols.length === 1);
-            case "pin_map":
-                // when we have a latitude and longitude a pin map is cool
-                return (data && this.hasLatitudeAndLongitudeColumns(data.cols));
-            case "line":
-            case "area":
-                // if we have 2x2 or more then that's enough to make a line/area chart
-                return (data && data.rows && data.rows.length > 1 && data.cols && data.cols.length > 1);
-            case "country":
-            case "state":
-                return (data && data.cols && data.cols.length > 1 && data.cols[0].base_type === "TextField");
-            case "bar":
-            case "pie":
-            default:
-                // general check for charts is that they require 2 columns
-                return (data && data.cols && data.cols.length > 1);
-        }
-    },
-
     isChartDisplay: function(display) {
         return (display !== "table" && display !== "scalar");
     },
 
-    setDisplay: function(type) {
-        // notify our parent about our change
-        this.props.setDisplayFn(type);
-    },
-
-    setChartColor: function(color) {
-        // tell parent about our new color
-        this.props.setChartColorFn(color);
-    },
-
-    renderChartTypePicker: function() {
-        var tetherOptions = {
-            attachment: 'bottom center',
-            targetAttachment: 'top center',
-            targetOffset: '-15px 0'
-        };
-        var currentIconName = this.visualizationTypeNames[this.props.card.display].iconName;
-        var chartTypeButton = (
-                <div className="inline-block">
-                    <span className="ChartType-button text-brand text-brand-hover shadowed flex layout-centered">
-                        <Icon name={currentIconName} width="30px" height="30px"></Icon>
-                    </span>
-                </div>
-        );
-
-        var displayOptions = this.props.visualizationTypes.map((type, index) => {
-            var classes = cx({
-                'p2': true,
-                'flex': true,
-                'align-center': true,
-                'cursor-pointer': true,
-                'bg-brand-hover': true,
-                'text-white-hover': true,
-                'ChartType--selected': type === this.props.card.display,
-                'ChartType--notSensible': !this.isSensibleChartDisplay(type),
-            });
-            var displayName = this.visualizationTypeNames[type].displayName;
-            var iconName = this.visualizationTypeNames[type].iconName;
-            return (
-                <li className={classes} key={index} onClick={this.setDisplay.bind(null, type)}>
-                    <Icon name={iconName} width="24px" height="24px"/>
-                    <span className="ml1">{displayName}</span>
-                </li>
-            );
-        });
-        return (
-            <span className="flex align-center">
-                Visualize as
-                <PopoverWithTrigger className="PopoverBody PopoverBody--withArrow ChartType-popover"
-                                    tetherOptions={tetherOptions}
-                                    triggerElement={chartTypeButton}>
-                    <ul className="">
-                        {displayOptions}
-                    </ul>
-                </PopoverWithTrigger>
-            </span>
-        );
-    },
-
-    renderChartColorPicker: function() {
-        if (this.props.card.display === "line" || this.props.card.display === "area" || this.props.card.display === "bar") {
-            var colors = this.props.visualizationSettingsApi.getDefaultColorHarmony();
-            var colorItems = [];
-            for (var i=0; i < colors.length; i++) {
-                var color = colors[i];
-                var localStyles = {
-                    "backgroundColor": color
-                };
-
-                colorItems.push((
-                    <li key={i} className="CardSettings-colorBlock" style={localStyles} onClick={this.setChartColor.bind(null, color)}></li>
-                ));
-            }
-
-            var colorPickerButton = (
-                <a className="Button">
-                    Change color
-                </a>
-            );
-
-            var tetherOptions = {
-                attachment: 'middle left',
-                targetAttachment: 'middle right',
-                targetOffset: '0 12px'
-            };
-
-            return (
-                <PopoverWithTrigger className="PopoverBody"
-                                    tetherOptions={tetherOptions}
-                                    triggerElement={colorPickerButton}>
-                    <ol className="p1">
-                        {colorItems}
-                    </ol>
-                </PopoverWithTrigger>
-            );
-
-        } else {
-            return false;
-        }
-    },
 
     renderFooter: function(tableFootnote) {
         if (this.props.isObjectDetail) {
             // no footer on object detail
             return false;
         } else {
-            var vizControls;
-            if (this.props.result && this.props.result.error === undefined) {
-                vizControls = (
-                    <div className="flex">
-                        {this.renderChartTypePicker()}
-                        {this.renderChartColorPicker()}
-                    </div>
-                );
-            }
-
             return (
-                <div className="VisualizationSettings wrapper flex">
-                    {vizControls}
+                <div className="VisualizationFooter wrapper flex">
                     {tableFootnote}
                 </div>
             );
